Remove things that won't be necessary
